Remove abstract class
