+++ plugins/com.google.cloud.tools.eclipse.dataflow.core/src/com/google/cloud/tools/eclipse/dataflow/core/launcher/DataflowPipelineLaunchDelegate.java
@@ -62,7 +62,7 @@ import org.osgi.framework.FrameworkUtil;
  *
  * <p>The JavaLaunchDelegate is responsible for most of the launching.
  */
-public class DataflowPipelineLaunchDelegate extends ForwardingLaunchConfigurationDelegate {
+public class DataflowPipelineLaunchDelegate implements ILaunchConfigurationDelegate2 {
   private static final String ARGUMENT_FORMAT_STR = "--%s=%s";
 
   @VisibleForTesting
@@ -98,8 +98,8 @@ public class DataflowPipelineLaunchDelegate extends ForwardingLaunchConfiguratio
       DataflowDependencyManager dependencyManager,
       IWorkspaceRoot workspaceRoot,
       IGoogleLoginService loginService) {
-    this.delegate = javaLaunchDelegate;
-    this.optionsRetrieverFactory = optionsHierarchyFactory;
+    delegate = javaLaunchDelegate;
+    optionsRetrieverFactory = optionsHierarchyFactory;
     this.dependencyManager = dependencyManager;
     this.workspaceRoot = workspaceRoot;
     this.loginService = loginService;
@@ -247,7 +247,25 @@ public class DataflowPipelineLaunchDelegate extends ForwardingLaunchConfiguratio
   }
 
   @Override
-  protected ILaunchConfigurationDelegate2 delegate() {
-    return delegate;
+  public ILaunch getLaunch(ILaunchConfiguration configuration, String mode) throws CoreException {
+    return delegate.getLaunch(configuration, mode);
+  }
+
+  @Override
+  public boolean buildForLaunch(ILaunchConfiguration configuration, String mode,
+      IProgressMonitor monitor) throws CoreException {
+    return delegate.buildForLaunch(configuration, mode, monitor);
+  }
+
+  @Override
+  public boolean finalLaunchCheck(ILaunchConfiguration configuration, String mode,
+      IProgressMonitor monitor) throws CoreException {
+    return delegate.finalLaunchCheck(configuration, mode, monitor);
+  }
+
+  @Override
+  public boolean preLaunchCheck(ILaunchConfiguration configuration, String mode,
+      IProgressMonitor monitor) throws CoreException {
+    return delegate.preLaunchCheck(configuration, mode, monitor);
   }
 }
